FastAPI/coursera-rest-api/pyproject.toml
[project]
name = "coursera-rest-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "pytest>=9.0.1",
    "httpx>=0.28.1",
    "databases>=0.9.0",
    "python-dotenv>=1.2.1",
    "pydantic-settings>=2.12.0",
    "aiosqlite>=0.21.0",
    "rich>=14.2.0",
]

[tool.black]
target-version = ["py312"]

